******************************************* ***** database need to be destroyed ! ***** *******************************************
* database : acquisition and acquisitionSummary : remove id column (not need and avoid declare an index, so data size is now half)
* database : acquisitionSummary : computations are done for each data insert (may change in future)
* plugin : RfxLan : SensorBasic and ControlBasic (in progress, works with our equipments)
* plugin : OpenZWave : fix mismatch type warnings

+++ sources/plugins/ZWave/OpenZWaveNodeCapacity.cpp
@@ -7,7 +7,7 @@
 #include "CommandClasses.h"
 #include "OpenZWaveNodeKeywordFactory.h"
 
-COpenZWaveNodeCapacity::COpenZWaveNodeCapacity(const int homeId, const int nodeId, const ECommandClass & classIdentifier)
+COpenZWaveNodeCapacity::COpenZWaveNodeCapacity(const uint32 homeId, const uint8 nodeId, const ECommandClass & classIdentifier)
    :m_classIdentifier(classIdentifier), m_class(OpenZWave::CommandClasses::CreateCommandClass(classIdentifier, homeId, nodeId))
 {
